+++ docutils/HISTORY.rst
@@ -26,6 +26,14 @@ Release 0.22rc3 (unpublished)
   - New transform `CitationReferences`. Marks citation_references
     as resolved if BibTeX is used by the backend (LaTeX).
 
+* docutils/parsers/rst/states.py
+
+  - Warn about duplicate name in references with embedded internal targets.
+
+* docutils/writers/manpage.py
+
+  - Do not drop text of internal targets.
+
 
 Release 0.22rc2 (2025-05-22)
 ============================
@@ -256,7 +264,6 @@ Release 0.22rc1 (2025-05-06)
   - Print Docutils version in header comment (feature-request #105).
   - Stop converting text to full capitals (bug #481).
   - Fix reference output (bug #497).
-  - Do not drop text of internal targets.
   - Use macros .UR/.UE for hyperlink references unless the new
     configuration setting text_references_ is True.
     The current default is True (text references), it will change